more efficient lock in check_old()

+++ trunk/mysql_conn.php
@@ -13,7 +13,15 @@ require_once 'Mailer.php';
  * Locks the table books to avoid repeated mail sending.
  */
 function check_old() {
-	mysql_query('lock table books write');
+	/* checking if antoher thread holds the lock */
+	$is_free_result = mysql_query('select is_free_lock("check_old")');
+	$is_free_row = mysql_fetch_row($is_free_result);
+	if ($is_free_row[0] == 0) return;
+	/* trying to get the lock */
+	$get_lock_result = mysql_query('select get_lock("check_old", 600)');
+	$get_lock_row = mysql_fetch_row($get_lock_result);
+	if ($get_lock_row[0] == 0) return;
+	/* now we habe the lock */
 	$query = 'delete from books where expired < now()';
 	mysql_query($query);
 	$query = 'select id,auth_key,mail,author,title,price,description from books where expired is null and expires < now()';
@@ -26,7 +34,7 @@ function check_old() {
 		$query = 'update books set expired = date_add(now(), interval 10 day) where id="'.$book['id'].'"';
 		mysql_query($query);
 	}
-	mysql_query('unlock tables');
+	mysql_query('select release_lock("check_old")');
 }
 
 check_old();
